Refactor to use `cargo bench` instead
